Rework test data and imports
