+++ packages/statemanager/tests/testdata/providerData/blocks/block0x7a11f.json
@@ -1,0 +1,22 @@
+{
+  "difficulty": "0x6dbe4d4bc9e",
+  "extraData": "0xd783010301844765746887676f312e342e32856c696e7578",
+  "gasLimit": "0x2fefd8",
+  "gasUsed": "0x0",
+  "hash": "0x794a1bef434928ce3aadd2f5eced2bf72ac714a30e9e4ab5965d7d9760300d84",
+  "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+  "miner": "0x52bc44d5378309ee2abf1539bf71de1b7d7be3b5",
+  "mixHash": "0x526746825a8fe8d43499f77d3d75536e7de0037c6440b4a5007fd0b2a1eaf5df",
+  "nonce": "0xe7ff3fe0cd3ac07b",
+  "number": "0x7a11f",
+  "parentHash": "0x18f9c888eed1f0b050052aa847dd8fd3962c3d4349ffa0059b1977fa3667342a",
+  "receiptsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
+  "sha3Uncles": "0x5a996aa4b940ff2893c07f98bd687f24e921d838a858d670d15aa189685fa131",
+  "size": "0x437",
+  "stateRoot": "0xce71e3f2f4edbb2c02b0bb1612f54c131b9bb1ad768aacf96afc2701d50062b1",
+  "timestamp": "0x563ceade",
+  "totalDifficulty": "0x2562baaf64c491c3",
+  "transactions": [],
+  "transactionsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
+  "uncles": ["0x8b2d2080fa241bfd0a182f2196ee605df68a5b2c9c89df3a070e1e6ea1c6ef62"]
+}
