+++ packages/statemanager/tests/testdata/providerData/blocks/block0x7a120.json
@@ -1,0 +1,22 @@
+{
+  "difficulty": "0x6dcc051573d",
+  "extraData": "0xd783010203844765746887676f312e352e31856c696e7578",
+  "gasLimit": "0x2fefd8",
+  "gasUsed": "0x0",
+  "hash": "0xac8e95f7483f7131261bcc0a70873f8236c27444c940defc677f74f281220193",
+  "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+  "miner": "0x580992b51e3925e23280efb93d3047c82f17e038",
+  "mixHash": "0x2e88e14c1a8cdf00e6252bd7303919f07b1fbe84d3767cc39e56fbd4c7b92204",
+  "nonce": "0x8901a71893cb3488",
+  "number": "0x7a120",
+  "parentHash": "0x794a1bef434928ce3aadd2f5eced2bf72ac714a30e9e4ab5965d7d9760300d84",
+  "receiptsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
+  "sha3Uncles": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347",
+  "size": "0x21d",
+  "stateRoot": "0xb2bcfa2ffe869085c84a976435f1581a7a0eb7af64bafcbbda710661016aa3ab",
+  "timestamp": "0x563ceae1",
+  "totalDifficulty": "0x2562c18c2515e900",
+  "transactions": [],
+  "transactionsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
+  "uncles": []
+}
